requirements.txt
vtk
Pillow
streamlit
opencv-python-headless
mediapipe
numpy
keras==2.15.0
streamlit-webrtc
av
tensorflow====2.15.0
twilio